ATO-398: Move logic from StorageTokenService to TokenService and delete StorageTokenService
